+++ packages/x-components/src/composables/index.ts
@@ -1,3 +1,5 @@
 export * from './create-use-device.composable';
 export * from './use-$x';
 export * from './use-register-x-module';
+export * from './use-on-display';
+