+++ packages/x-components/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.0-alpha.13](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.12...@empathyco/x-components@4.1.0-alpha.13) (2024-01-25)
+
+
+### Features
+
+* **queries-preview:** allow queries preview with same query but different filters or params (#1392) ([4c1f63a](https://github.com/empathyco/x/commit/4c1f63aaf4e9fc75d3b8298b2ea41c1fc207940a))
+
+
+
+## [4.1.0-alpha.12](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.11...@empathyco/x-components@4.1.0-alpha.12) (2024-01-25)
+
+
+### Features
+
+* **components:** composable and component to fire callbacks when an element appears on viewport (#1391) ([c463352](https://github.com/empathyco/x/commit/c46335243c74c604e6f8168c158461edff9d4eb5))
+
+
+
 ## [4.1.0-alpha.11](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.10...@empathyco/x-components@4.1.0-alpha.11) (2024-01-22)